+++ TaskListPlugin/DockPanes/TaskListPane.cs
@@ -26,17 +26,18 @@ namespace SphereStudio.UI
             iconImageList.Images.Add("not", Resources.lightbulb);
             iconImageList.Images.Add("done", Resources.lightbulb_off);
             taskListView.SmallImageList = iconImageList;
+            taskListView.AlwaysGroupByColumn = olvColumn3;
 
             olvColumn1.ImageGetter = delegate (object rowObject)
             {
-                var entry = (TaskListItem)rowObject;
-                return entry.IsFinished ? "done" : "not";
+                var taskListItem = (TaskListItem)rowObject;
+                return taskListItem.Finished ? "done" : "not";
             };
 
-            var taskTypeNames = Enum.GetNames(typeof(TaskType));
+            var taskCategoryNames = Enum.GetNames(typeof(TaskCategory));
             var priorityNames = Enum.GetNames(typeof(TaskPriority));
-            foreach (var name in taskTypeNames)
-                setTypeMenuItem.DropDownItems.Add(name, Resources.lightbulb, taskTypeMenuItem_Click);
+            foreach (var name in taskCategoryNames)
+                setCategoryMenuItem.DropDownItems.Add(name, Resources.lightbulb, taskCategoryMenuItem_Click);
             foreach (var name in priorityNames)
                 setPriorityMenuItem.DropDownItems.Add(name, Resources.resultset_none, taskPriorityMenuItem_Click);
         }
@@ -80,10 +81,10 @@ namespace SphereStudio.UI
                 {
                     var taskListItem = new TaskListItem()
                     {
-                        IsFinished = reader.ReadBoolean(),
+                        Finished = reader.ReadBoolean(),
                         Name = reader.ReadString(),
                         Priority = (TaskPriority)reader.ReadInt32(),
-                        Type = (TaskType)reader.ReadInt32()
+                        Category = (TaskCategory)reader.ReadInt32()
                     };
                     taskList.Add(taskListItem);
                 }
@@ -100,10 +101,10 @@ namespace SphereStudio.UI
                 fileWriter.Write(taskListView.GetItemCount());
                 foreach (TaskListItem task in taskListView.Objects)
                 {
-                    fileWriter.Write(task.IsFinished);
+                    fileWriter.Write(task.Finished);
                     fileWriter.Write(task.Name);
                     fileWriter.Write((int)task.Priority);
-                    fileWriter.Write((int)task.Type);
+                    fileWriter.Write((int)task.Category);
                 }
                 fileWriter.Flush();
             }
@@ -139,9 +140,11 @@ namespace SphereStudio.UI
             increasePriorityToolButton.Enabled = haveSelection;
             removeTaskToolButton.Enabled = haveSelection;
 
-            removeTaskMenuItem.Enabled = haveSelection;
+            decreasePriorityMenuItem.Enabled = haveSelection;
+            deleteTaskMenuItem.Enabled = haveSelection;
+            increasePriorityMenuItem.Enabled = haveSelection;
             setPriorityMenuItem.Enabled = haveSelection;
-            setTypeMenuItem.Enabled = haveSelection;
+            setCategoryMenuItem.Enabled = haveSelection;
         }
 
         private void taskListView_FormatCell(object sender, FormatCellEventArgs e)
@@ -149,7 +152,7 @@ namespace SphereStudio.UI
             var taskListItem = e.Model as TaskListItem;
             if (e.ColumnIndex == olvColumn1.Index && taskListItem != null)
             {
-                var fontStyle = taskListItem.IsFinished ? FontStyle.Strikeout : FontStyle.Regular;
+                var fontStyle = taskListItem.Finished ? FontStyle.Strikeout : FontStyle.Regular;
                 e.SubItem.Font = new Font(e.SubItem.Font, fontStyle);
             }
         }
@@ -183,43 +186,48 @@ namespace SphereStudio.UI
             refreshMenuCommands();
         }
 
-        private void decreasePriorityToolButton_Click(object sender, EventArgs e)
+        private void addTaskMenuItem_Click(object sender, EventArgs e)
+        {
+            taskListView.AddObject(new TaskListItem("New Task"));
+        }
+
+        private void deleteAllTasksMenuItem_Click(object sender, EventArgs e)
+        {
+            var projectName = PluginManager.Core.Project.Name;
+            var dialogResult = MessageBox.Show(
+                $"Are you sure you want to delete all tasks, including incomplete tasks, for the project {projectName}?",
+                "Delete All Tasks", MessageBoxButtons.YesNo, MessageBoxIcon.Warning);
+            if (dialogResult == DialogResult.Yes)
+                Clear();
+        }
+
+        private void decreasePriorityMenuItem_Click(object sender, EventArgs e)
         {
             foreach (TaskListItem task in taskListView.SelectedObjects)
                 task.DecreasePriority();
             taskListView.RefreshSelectedObjects();
         }
 
-        private void increasePriorityToolButton_Click(object sender, EventArgs e)
+        private void deleteTaskMenuItem_Click(object sender, EventArgs e)
+        {
+            foreach (TaskListItem task in taskListView.SelectedObjects)
+                taskListView.RemoveObject(task);
+        }
+
+        private void increasePriorityMenuItem_Click(object sender, EventArgs e)
         {
             foreach (TaskListItem task in taskListView.SelectedObjects)
                 task.IncreasePriority();
             taskListView.RefreshSelectedObjects();
         }
 
-        private void addTaskMenuItem_Click(object sender, EventArgs e)
-        {
-            taskListView.AddObject(new TaskListItem("New Task"));
-        }
-
-        private void removeAllTasksMenuItem_Click(object sender, EventArgs e)
-        {
-            Clear();
-        }
-
-        private void removeCompletedTasksMenuItem_Click(object sender, EventArgs e)
+        private void pruneTasksMenuItem_Click(object sender, EventArgs e)
         {
             var removedItems = taskListView.Objects
                 .Cast<TaskListItem>()
-                .Where(entry => entry.IsFinished)
+                .Where(entry => entry.Finished)
                 .ToArray();
             taskListView.RemoveObjects(removedItems);
-        }
-
-        private void removeTaskMenuItem_Click(object sender, EventArgs e)
-        {
-            foreach (TaskListItem task in taskListView.SelectedObjects)
-                taskListView.RemoveObject(task);
         }
 
         private void taskPriorityMenuItem_Click(object sender, EventArgs e)
@@ -232,12 +240,12 @@ namespace SphereStudio.UI
             }
         }
 
-        private void taskTypeMenuItem_Click(object sender, EventArgs e)
+        private void taskCategoryMenuItem_Click(object sender, EventArgs e)
         {
             foreach (TaskListItem taskListItem in taskListView.SelectedObjects)
             {
-                var index = setTypeMenuItem.DropDownItems.IndexOf((ToolStripItem)sender);
-                taskListItem.Type = (TaskType)index;
+                var index = setCategoryMenuItem.DropDownItems.IndexOf((ToolStripItem)sender);
+                taskListItem.Category = (TaskCategory)index;
                 taskListView.RefreshObject(taskListItem);
             }
         }