+++ CHANGELOG.md
@@ -5,6 +5,7 @@ v2.3.2 - TBD
 ------------
 
 * Improves readability of Task List entries.
+* Fixes a bug where items in the Task List pane couldn't be marked completed.
 * Fixes a bug where clicking Cancel in the "unsaved changes" prompt didn't
   stop the IDE from closing, causing the changes to be lost.